+++ stats/wagon/wagon_aux.cc
@@ -459,7 +459,7 @@ float WImpurity::vector_impurity()
     /* simple distance */
     for (j=0; j<wgn_VertexFeats.num_channels(); j++)
     {
-        if (wgn_VertexFeats.a(0L,j) > 0.0)
+        if (wgn_VertexFeats.a(static_cast<ssize_t>(0),j) > 0.0)
         {
             b.reset();
             for (pp=members.head(), countpp=member_counts.head(); pp != 0; pp=pp->next(), countpp=countpp->next())
@@ -589,8 +589,8 @@ float WImpurity::trajectory_impurity()
         i = members.item(pp);
         for (q=0; q<wgn_UnitTrack.a(i,1); q++)
         {
-            ni = wgn_UnitTrack.a(i,0L)+q;
-            if (wgn_VertexTrack.a(ni,0L) == -1.0)
+            ni = wgn_UnitTrack.a(i,0)+q;
+            if (wgn_VertexTrack.a(ni,0) == -1.0)
             {
                 l1ss += q;
                 ola = 1;
@@ -623,13 +623,13 @@ float WImpurity::trajectory_impurity()
         {   /* for each unit */
             i = members.item(pp);
             m = (float)wgn_UnitTrack.a(i,1L)/(float)l; /* find interpolation */
-            s = wgn_UnitTrack.a(i,0L); /* start point */
+            s = wgn_UnitTrack.a(i,0); /* start point */
             for (ti=0,n=0.0; ti<l; ti++,n+=m)
             {
                 ni = (int)n;  // hmm floor or nint ??
                 for (j=0; j<wgn_VertexFeats.num_channels(); j++)
                 {
-                    if (wgn_VertexFeats.a(0L,j) > 0.0)
+                    if (wgn_VertexFeats.a(static_cast<ssize_t>(0),j) > 0.0)
                         trajectory[ti][j] += wgn_VertexTrack.a(s+ni,j);
                 }
             }
@@ -640,7 +640,7 @@ float WImpurity::trajectory_impurity()
         for (ti=0; ti<l; ti++)
             for (j=0; j<wgn_VertexFeats.num_channels(); j++)
             {
-                if (wgn_VertexFeats.a(0L,j) > 0.0)
+                if (wgn_VertexFeats.a(static_cast<ssize_t>(0),j) > 0.0)
                     stdss += trajectory[ti][j].stddev();
             }
 
@@ -662,9 +662,9 @@ float WImpurity::trajectory_impurity()
         {   /* for each unit */
             i = members.item(pp);
             s1l = 0;
-            s = wgn_UnitTrack.a(i,0L); /* start point */
+            s = wgn_UnitTrack.a(i,0); /* start point */
             for (q=0; q<wgn_UnitTrack.a(i,1); q++)
-                if (wgn_VertexTrack.a(s+q,0L) == -1.0)
+                if (wgn_VertexTrack.a(s+q,0) == -1.0)
                 {
                     s1l = q; /* printf("awb q is -1 at %d\n",q); */
                     break;
@@ -677,12 +677,12 @@ float WImpurity::trajectory_impurity()
             {
                 ni = s + (((int)n < s1l) ? (int)n : s1l - 1);
                 for (j=0; j<wgn_VertexFeats.num_channels(); j++)
-                    if (wgn_VertexFeats.a(0L,j) > 0.0)
+                    if (wgn_VertexFeats.a(static_cast<ssize_t>(0),j) > 0.0)
                         trajectory[ti][j] += wgn_VertexTrack.a(ni,j);
             }
             ti = l1; /* do it explicitly in case s1l < 1 */
             for (j=0; j<wgn_VertexFeats.num_channels(); j++)
-                if (wgn_VertexFeats.a(0L,j) > 0.0)
+                if (wgn_VertexFeats.a(static_cast<ssize_t>(0),j) > 0.0)
                     trajectory[ti][j] += -1;
             /* Second half */
             s += s1l+1;
@@ -690,11 +690,11 @@ float WImpurity::trajectory_impurity()
             {
                 ni = s + (((int)n < s2l) ? (int)n : s2l - 1);
                 for (j=0; j<wgn_VertexFeats.num_channels(); j++)
-                    if (wgn_VertexFeats.a(0L,j) > 0.0)
+                    if (wgn_VertexFeats.a(static_cast<ssize_t>(0),j) > 0.0)
                         trajectory[ti][j] += wgn_VertexTrack.a(ni,j);
             }
             for (j=0; j<wgn_VertexFeats.num_channels(); j++)
-                if (wgn_VertexFeats.a(0L,j) > 0.0)
+                if (wgn_VertexFeats.a(static_cast<ssize_t>(0),j) > 0.0)
                     trajectory[ti][j] += -2;
         }
 
@@ -704,12 +704,12 @@ float WImpurity::trajectory_impurity()
         m = 1.0/(float)l1;
         for (w=0.0,ti=0; ti<l1; ti++,w+=m)
             for (j=0; j<wgn_VertexFeats.num_channels(); j++)
-                if (wgn_VertexFeats.a(0L,j) > 0.0)
+                if (wgn_VertexFeats.a(static_cast<ssize_t>(0),j) > 0.0)
                 stdss += trajectory[ti][j].stddev() * w;
         m = 1.0/(float)l2;
         for (w=1.0,ti++; ti<l-1; ti++,w-=m)
             for (j=0; j<wgn_VertexFeats.num_channels(); j++)
-                if (wgn_VertexFeats.a(0L,j) > 0.0)
+                if (wgn_VertexFeats.a(static_cast<ssize_t>(0),j) > 0.0)
                     stdss += trajectory[ti][j].stddev() * w;
     
         // This is sum of all stddev * samples
@@ -800,7 +800,7 @@ float WImpurity::ols_impurity()
     ols_test(Y,pred,cor,rmse);
     best_score = cor;
 
-    printf("Impurity OLS X(%ld,%ld) Y(%ld,%ld) %f, %f, %f\n",
+    printf("Impurity OLS X(%zd,%zd) Y(%zd,%zd) %f, %f, %f\n",
              X.num_rows(),X.num_columns(),Y.num_rows(),Y.num_columns(),
              rmse,cor,
              1-best_score);
@@ -1004,7 +1004,7 @@ ostream & operator <<(ostream &s, WImpurity &imp)
             cs = new EST_SuffStats [wgn_VertexTrack.num_channels()+1];
             
             for (j=0; j<wgn_VertexFeats.num_channels(); j++)
-                if (wgn_VertexFeats.a(0L,j) > 0.0)
+                if (wgn_VertexFeats.a(static_cast<ssize_t>(0),j) > 0.0)
                 {
                     cs[j].reset();
                     for (p=imp.members.head(); p != 0; p=p->next())
@@ -1016,7 +1016,7 @@ ostream & operator <<(ostream &s, WImpurity &imp)
             for (p=imp.members.head(); p != 0; p=p->next())
             {
                 for (x=0.0,j=0; j<wgn_VertexFeats.num_channels(); j++)
-                    if (wgn_VertexFeats.a(0L,j) > 0.0)
+                    if (wgn_VertexFeats.a(static_cast<ssize_t>(0),j) > 0.0)
                     {
                         d = (wgn_VertexTrack.a((ssize_t)imp.members.item(p),j)-cs[j].mean())
                             /* / cs[j].stddev() */ ; /* seems worse 061218 */
